Move `Server` into a lib
This allows us running a minimal binary in `src/main.rs` as well as
maintaining everything else within a conmon server library.
