+++ Cargo.toml
@@ -9,4 +9,3 @@ members = [
 lto = true
 opt-level = "z"
 codegen-units = 1
-
