+++ conmon-rs/server/Cargo.toml
@@ -3,6 +3,10 @@ name = "conmon-server"
 version = "0.1.0"
 edition = "2021"
 
+[[bin]]
+name = "conmon-server"
+path = "src/main.rs"
+
 [dependencies]
 anyhow = "1.0.45"
 capnp = "0.14.3"
@@ -18,4 +22,3 @@ simple_logger = "1.13.0"
 tokio = { version = "1.14.0", features = ["fs", "macros", "net", "process", "rt", "signal", "time"] }
 tokio-util = { version = "0.6.9", features = ["compat"] }
 nix = "0.23.0"
-
